Moves route files into dedicated directory
Organizes application structure by relocating route-related files
to a specific directory and updating import paths for consistency.
Improves maintainability and clarity of file organization.
