+++ src/app/(routes)/dashboard/page.tsx
