+++ src/app/(routes)/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
-import "./globals.css";
-import AppProviders from "../providers/app_providers";
+import "../globals.css";
+import AppProviders from "../../providers/app_providers";
 
 const geistSans = Geist({
   variable: "--font-geist-sans",
